factorize html generator for error responses
