+++ shared/core-utils/miscs/http-error-codes.ts
@@ -257,6 +257,8 @@ export enum HttpStatusCode {
   EXPECTATION_FAILED_417 = 417,
 
   /**
+   * Official Documentation @ https://tools.ietf.org/html/rfc2324
+   *
    * This code was defined in 1998 as one of the traditional IETF April Fools' jokes, in RFC 2324, Hyper Text Coffee Pot Control Protocol,
    * and is not expected to be implemented by actual HTTP servers. The RFC specifies this code should be returned by
    * teapots requested to brew coffee. This HTTP status is used as an Easter egg in some websites, including PeerTube instances ;-).
